auth-form moved to components folder
